requirements.txt
weaviate-client
chembl_webresource_client
sentence-transformers
pandas
tqdm